some fixes for --strong warnings I didn't go all the way, in case we want to change some of these. Thoughts?
R=leafp@google.com

Review URL: https://codereview.chromium.org/1245013002 .

+++ pkg/dev_compiler/lib/src/report.dart
@@ -204,7 +204,7 @@ class _Table {
       while (abbreviations[headerName] != null) headerName = "$headerName'";
       abbreviations[headerName] = name;
     }
-    widths.add(max(5, headerName.length + 1));
+    widths.add(max(5, headerName.length + 1) as int);
     header.add(headerName);
     _totalColumns++;
   }
